HTTPApi:				1- Changed the returned closure value to return a generic type. Also removed the array as I didn't want to loop through the results but rather just display them all. Also added an error return for use during errors.
Added new Details view controller, coordinator and details view for configuration.

SearchResultsView:		1 - Moved from Mock data to calling the API.

SearchResultsCoor		3- Changed the param to match the passed in JSON.
					7- Details coor setup
					8/11- Made changes to data passed into to setup search results view
					11- Added new fetch details method.
					12- Delegate method for later use.
SearchResultsColDel	1- Delegate for later use to send the selected movie back to the coordinator for API use. This allows for reuse and abstraction of the delegate.